Refactor `ServerState` to move the `initialize_global_sync_workspace` function into the struct and update all related calls.
